% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yxl_tad.R
\name{yxl_tad}
\alias{yxl_tad}
\title{Use tad as an external data browser. tad needs to be installed
independently}
\usage{
yxl_tad(df)
}
\arguments{
\item{df}{the data.frame to be exported.}
}
\description{
Use tad as an external data browser. tad needs to be installed
independently
}
